Completed all the backend
number buttons now work to write their number
to the calculator. Delete button deletes the last
number added. Clear button clears the active number,
if pressed when the active number is already clear
clears te memory of the calculator. Equals computes
the result of the operation. The operation button
save the active number to memory as well as the operation
to be executed on it.

[Ticket: X]